+++ Src/Commands/AddDataCommand.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
+using System.Reflection;
 
 namespace Ldbc.Commands
 {
@@ -26,10 +27,14 @@ namespace Ldbc.Commands
 		public SwitchParameter Bulk { get; set; }
 		List<object> _bulk;
 
+		static MethodInfo _RemoveDocId; //rk work around
+
 		protected override void BeginProcessing()
 		{
 			if (Bulk)
 				_bulk = new List<object>();
+			else
+				_RemoveDocId = typeof(LiteCollection<BsonDocument>).GetMethod("RemoveDocId", BindingFlags.NonPublic | BindingFlags.Instance, null, new Type[] { typeof(BsonDocument) }, null);
 		}
 
 		protected override void ProcessRecord()
@@ -46,6 +51,8 @@ namespace Ldbc.Commands
 			try
 			{
 				var document = Actor.ToBsonDocument(InputObject);
+				//rk work around
+				_RemoveDocId.Invoke(Collection, new object[] { document });
 				var id = Collection.Insert(document);
 				if (Result)
 					WriteObject(Actor.ToObject(id));