+++ Src/Commands/UpdateDataCommand.cs
@@ -14,8 +14,8 @@ namespace Ldbc.Commands
 		public ILiteCollection<BsonDocument> Collection { get; set; }
 
 		[Parameter(Position = 1, Mandatory = true)]
-		public object Filter { set { _Filter = Expression.Create(value); } }
-		Expression _Filter;
+		public object Where { set { _Where = Expression.Create(value); } }
+		Expression _Where;
 
 		[Parameter(Position = 2, Mandatory = true)]
 		public object Update { set { _Update = Expression.Create(value); } }
@@ -26,7 +26,7 @@ namespace Ldbc.Commands
 
 		protected override void BeginProcessing()
 		{
-			var count = Collection.UpdateMany(_Update.BsonExpression, _Filter.BsonExpression);
+			var count = Collection.UpdateMany(_Update.BsonExpression, _Where.BsonExpression);
 			if (Result)
 				WriteObject(count);
 		}