+++ Tests/Class.test.ps1
@@ -52,3 +52,25 @@ task ClassExample {
 	equals $r[0].Id 1
 	equals $r[0].Location.Line 11
 }
+
+#! https://github.com/mbdavid/LiteDB/issues/1483
+#! auto id is currently not working as expected
+#! when resolved, remove my work around
+task TODO {
+	Use-LiteDatabase :memory: {
+		$test = Get-LiteCollection Test Int32
+		try {
+			# InsertBulk is fine
+			Add-LiteData $test @{_id = 0} -Bulk
+
+			# Insert has issue
+			$d = [LiteDB.BsonDocument]::new()
+			$d["_id"] = 0
+			$test.Insert($d)
+			throw
+		}
+		catch {
+			assert ("$_" -like '*Object reference not set to an instance of an object.*')
+		}
+	}
+}